Fixed warning and other small perfections

+++ 5003_Task2a_CashPoint_Skeleton_13-4/UserInterface.cpp
@@ -231,7 +231,7 @@ double UserInterface::readInDepositAmount() const {
 int UserInterface::readInNumberOfTransactions() const 
 {
 	cout << "\nNUMBER OF TRANSACTIONS : ";
-	return ( readInPositiveAmount());
+	return ( static_cast<int>(readInPositiveAmount()));
 }
 bool UserInterface::readInConfirmDeletion() const
 {
